+++ app/[lang]/layout.tsx
@@ -1,4 +1,3 @@
-import type { Metadata } from 'next';
 import { Montserrat, Raleway } from 'next/font/google';
 
 import { Locale, i18n } from '@/i18n.config';
@@ -18,39 +17,39 @@ const raleway = Raleway({
   variable: '--font-raleway',
 });
 
-export async function generateMetadata({
-  params: { lang },
-}: {
-  params: { lang: Locale };
-}): Promise<Metadata> {
-  const {
-    common: { meta },
-  } = await getDictionary(lang);
+// export async function generateMetadata({
+//   params: { lang },
+// }: {
+//   params: { lang: Locale };
+// }): Promise<Metadata> {
+//   const {
+//     common: { meta },
+//   } = await getDictionary(lang);
 
-  const { metadata, twitter, openGraph, icons } = meta;
-  const { title, description, keywords, manifest } = metadata;
+//   const { metadata, twitter, openGraph, icons } = meta;
+//   const { title, description, keywords, manifest } = metadata;
 
-  //TODO: delete localhost:3000 / create .env.local
-  const baseUrl = process.env.BASE_URL || 'http://localhost:3000/';
+//   //TODO: delete localhost:3000 / create .env.local
+//   const baseUrl = process.env.BASE_URL || 'http://localhost:3000/';
 
-  return {
-    title,
-    description,
-    metadataBase: new URL(baseUrl),
-    alternates: {
-      canonical: `${baseUrl}${lang}`,
-      languages: {
-        'uk-UA': '/uk',
-        'en-US': '/en',
-      },
-    },
-    keywords,
-    manifest,
-    twitter,
-    openGraph: { ...openGraph, url: `${baseUrl}${lang}` },
-    icons,
-  };
-}
+//   return {
+//     title,
+//     description,
+//     metadataBase: new URL(baseUrl),
+//     alternates: {
+//       canonical: `${baseUrl}${lang}`,
+//       languages: {
+//         'uk-UA': '/uk',
+//         'en-US': '/en',
+//       },
+//     },
+//     keywords,
+//     manifest,
+//     twitter,
+//     openGraph: { ...openGraph, url: `${baseUrl}${lang}` },
+//     icons,
+//   };
+// }
 
 export async function generateStaticParams() {
   return i18n.locales.map(locale => ({ lang: locale }));
