+++ dictionaries/en/meta/home.json
@@ -1,0 +1,7 @@
+{
+  "metadataHome": {
+    "title": "13 Candle Bar",
+    "description": "Family workshop of aromas and coziness",
+    "keywords": "Aroma Candles,  Candles in planters, Aroma diffusers, Aroma sprays, Incense"
+  }
+}
