+++ dictionaries/uk/common.json
@@ -105,8 +105,8 @@
         "images": [
           {
             "url": "/meta/ogp-image.jpg",
-            "width": 1280,
-            "height": 640,
+            "width": 1017,
+            "height": 509,
             "alt": "13 Candle Bar. Сімейна майстерня ароматів і затишку"
           }
         ]
