+++ dictionaries/uk/meta/home.json
@@ -1,0 +1,7 @@
+{
+  "metadataHome": {
+    "title": "13 Candle Bar ",
+    "description": "Сімейна майстерня ароматів і затишку",
+    "keywords": "Аромасвічки, Свічки в кашпо, Аромадифузори, Аромаспреї, Пахощі"
+  }
+}
